backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Django==4.2.5
django-auditlog==2.3.0
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
fonttools==4.43.1
kiwisolver==1.4.5
kombu==5.3.2
matplotlib==3.8.0
numpy==1.26.0
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
shapely==2.0.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
